requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.54.4
transformers==4.46.0
torch==2.5.1
torchvision==0.20.1
accelerate==1.1.1
Pillow==10.4.0
requests>=2.32.0
numpy>=1.26.0,<2.0.0
protobuf>=3.20.0,<5.0.0